requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
Markdown==2.6.8
pkg-resources==0.0.0
PyJWT==1.5.2
pytz==2017.2
